% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{ComputeLinearModel}
\alias{ComputeLinearModel}
\title{Fits linear regression where APC Variables are numeric}
\usage{
ComputeLinearModel(data, predictors)
}
\arguments{
\item{data}{output from 'GetDataCohortTableCustom'}

\item{predictors}{c(character) e.g. c("age", "period")}
}
\value{
named list: (1) the data used to fit the model and (2) the model 
object returned by lm()
}
\description{
Fits linear regression where APC Variables are numeric
}
