% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{ComputeRecencyFrequency}
\alias{ComputeRecencyFrequency}
\title{Computes for each customer
    (1) Recency := timestamp of last order
    (2) Frequency := average number of purchases per month}
\usage{
ComputeRecencyFrequency(translog)
}
\value{
data.table with columns: "custId", "frequency", "recency"
}
\description{
Computes for each customer
    (1) Recency := timestamp of last order
    (2) Frequency := average number of purchases per month
}
